Move ComplianceOperatorInfo pipeline
