+++ central/complianceoperator/v2/pipelines/complianceoperatorinfo/pipeline.go
