+++ central/complianceoperator/v2/pipelines/complianceoperatorinfo/pipeline_test.go
